split placeholder additions
