+++ openfisca_us/variables/gov/states/md/tax/income/additions/other_additions/md_oil_percentage_depletion_allowance_addition.py
@@ -1,0 +1,13 @@
+## md_oil_percentage_depletion_allowance_addition.py
+from openfisca_us.model_api import *
+
+class md_oil_percentage_depletion_allowance_addition(Variable):
+    # f. Oil percentage depletion allowance claimed under IRC Section 613.
+    value_type = float
+    entity = TaxUnit
+    label = "MD Oil Percentage Depletion Allowance"
+    documentation = (
+        "Oil percentage depletion allowance claimed under IRC Section 613"
+    )
+    unit = USD
+    definition_period = YEAR
